+++ R/textbox.R
@@ -76,6 +76,10 @@ grid.textbox <- function(...) {
 }
 
 get_hw <- function(x) {
+  if (has_markup(x$label)) {
+    return(get_hw_formatted(x))
+  }
+
   t <- textGrob(label = x$label, gp = x$txt_gp)
   # Add padding
   padding <- unit(1 * ifelse(is.null(x$txt_gp$cex), 1,
@@ -85,6 +89,41 @@ get_hw <- function(x) {
   height <- convertHeight(grobHeight(t), "char") + padding
   # width <- grobWidth(t) + padding
   width <- convertWidth(grobWidth(t), "char") + padding
+
+  list(width = width, height = height)
+}
+
+# Measure width/height for formatted (markup) text
+get_hw_formatted <- function(x) {
+  segments <- parse_markup(x$label)
+  lines    <- split_segments_by_newline(segments)
+
+  cex     <- ifelse(is.null(x$txt_gp$cex), 1, x$txt_gp$cex)
+  padding <- unit(1 * cex, "char")
+
+  # Width of each line in inches
+  line_widths <- sapply(lines, function(line_segs) {
+    sum(sapply(line_segs, function(seg) {
+      gp <- segment_gpar(seg$style, x$txt_gp)
+      t  <- textGrob(label = seg$text, gp = gp)
+      convertWidth(grobWidth(t), "inches", valueOnly = TRUE)
+    }))
+  })
+
+  width <- convertWidth(unit(max(line_widths), "inches"), "char") + padding
+
+  # Height from a dummy multi-line textGrob (correct line spacing)
+  n_lines    <- length(lines)
+  dummy_text <- paste(rep("Xg", n_lines), collapse = "\n")
+  dummy      <- textGrob(label = dummy_text, gp = x$txt_gp)
+  height     <- convertHeight(grobHeight(dummy), "char") + padding
+
+  # Extra room for super/subscript offsets
+  has_ss <- any(sapply(segments, function(s)
+    s$style %in% c("superscript", "subscript")))
+  if (has_ss) {
+    height <- height + unit(0.3 * cex, "char")
+  }
 
   list(width = width, height = height)
 }
@@ -114,21 +153,26 @@ makeContent.textbox <- function(x) {
   # Add padding
   padding <- unit(3 * ifelse(is.null(x$txt_gp$cex), 1, x$txt_gp$cex), "mm")
 
-  # Align text
-  tx_x <- switch(x$just,
-    "right" = unit(1, "npc") - 0.5 * padding,
-    "left"  = 0.5 * padding,
-    "center" = unit(.5, "npc")
-  )
+  if (has_markup(x$label)) {
+    text_grobs <- make_formatted_content(x)
+  } else {
+    # Align text
+    tx_x <- switch(x$just,
+      "right" = unit(1, "npc") - 0.5 * padding,
+      "left"  = 0.5 * padding,
+      "center" = unit(.5, "npc")
+    )
 
-  t <- textGrob(
-    label = x$label,
-    x = tx_x,
-    y = 0.5,
-    just = x$just,
-    gp = x$txt_gp,
-    name = paste0(x$name, ".text")
-  )
+    t <- textGrob(
+      label = x$label,
+      x = tx_x,
+      y = 0.5,
+      just = x$just,
+      gp = x$txt_gp,
+      name = paste0(x$name, ".text")
+    )
+    text_grobs <- gList(t)
+  }
 
   hw <- get_hw(x)
 
@@ -141,7 +185,86 @@ makeContent.textbox <- function(x) {
     name = paste0(x$name, ".box")
   ))
 
-  setChildren(x, gList(rr, t))
+  setChildren(x, do.call(gList, c(list(rr), as.list(text_grobs))))
+}
+
+# Render formatted (markup) text as positioned grid grobs
+make_formatted_content <- function(x) {
+  segments <- parse_markup(x$label)
+  lines    <- split_segments_by_newline(segments)
+  n_lines  <- length(lines)
+
+  cex     <- ifelse(is.null(x$txt_gp$cex), 1, x$txt_gp$cex)
+  padding <- unit(3 * cex, "mm")
+
+  grob_list <- list()
+
+  for (i in seq_along(lines)) {
+    line_segs <- lines[[i]]
+
+    # Vertical position: evenly distribute lines top-to-bottom
+    y_frac <- 1 - (2 * i - 1) / (2 * n_lines)
+    y_base <- unit(y_frac, "npc")
+
+    # Measure each segment width (in inches, resolved now)
+    seg_info <- lapply(seq_along(line_segs), function(j) {
+      seg <- line_segs[[j]]
+      gp  <- segment_gpar(seg$style, x$txt_gp)
+      tg  <- textGrob(seg$text, gp = gp, just = "left")
+      w   <- convertWidth(grobWidth(tg), "inches", valueOnly = TRUE)
+      list(grob = tg, style = seg$style, width_in = w)
+    })
+
+    total_width_in <- sum(sapply(seg_info, "[[", "width_in"))
+
+    # Horizontal start position based on justification
+    start_x <- switch(x$just,
+      "center" = unit(0.5, "npc") - unit(total_width_in / 2, "inches"),
+      "left"   = padding * 0.5,
+      "right"  = unit(1, "npc") - padding * 0.5 - unit(total_width_in, "inches"),
+      unit(0.5, "npc") - unit(total_width_in / 2, "inches")
+    )
+
+    cum_x_in <- 0
+
+    for (j in seq_along(seg_info)) {
+      si <- seg_info[[j]]
+
+      # Vertical offset for super/subscript
+      y_off <- switch(si$style,
+        "superscript" = unit(0.4 * cex, "char"),
+        "subscript"   = unit(-0.3 * cex, "char"),
+        unit(0, "mm")
+      )
+
+      si$grob$x    <- start_x + unit(cum_x_in, "inches")
+      si$grob$y    <- y_base + y_off
+      si$grob$name <- sprintf("%s.text.%d.%d", x$name, i, j)
+
+      grob_list[[length(grob_list) + 1L]] <- si$grob
+
+      # Draw underline beneath the segment
+      if (si$style == "underline") {
+        seg_w <- unit(si$width_in, "inches")
+        ul_y  <- y_base - unit(0.55 * cex, "char")
+        ul_x0 <- start_x + unit(cum_x_in, "inches")
+        ul <- segmentsGrob(
+          x0 = ul_x0, x1 = ul_x0 + seg_w,
+          y0 = ul_y,  y1 = ul_y,
+          gp = gpar(
+            col = if (!is.null(x$txt_gp$col)) x$txt_gp$col else "black",
+            lwd = max(0.5, 0.8 * cex)
+          ),
+          name = sprintf("%s.ul.%d.%d", x$name, i, j)
+        )
+        grob_list[[length(grob_list) + 1L]] <- ul
+      }
+
+      cum_x_in <- cum_x_in + si$width_in
+    }
+  }
+
+  do.call(gList, grob_list)
 }
 
 #' @export